Refactor: keep model goldens and tooling in the model tree
- revert the golden package additions: the three full-net comparators
  move into models/deepseek_v4_pro/utils/golden_fwd.py next to their
  only consumer, still composing the public ratio_allclose primitive;
  golden/runner.py keeps the generic effective-scalar pass-through
- weights_flash.py and golden_fwd.py move to
  models/deepseek_v4_pro/utils/ (weights_flash bootstraps sys.path so
  its conversion CLI still runs as a script)
- the row-count-agnostic hc_head golden and the all-ranks lm_head
  replay move into hc_head.py and lm_head.py as golden_hc_head_rows
  and golden_lm_head_all_ranks(n_ranks=...); utils/golden_fwd.py keeps
  only the layer-schedule orchestration and the compare-fn builder
- drop the tests/models regression files; the variant-agnostic Q_LORA
  floor fix stays in tests/golden/test_validation.py
- docs point at the moved conversion CLI
