+++ golden/__init__.py
@@ -46,12 +46,9 @@ from .runner import RunResult, run, run_jit
 from .spec import ScalarSpec, TensorSpec
 from .validation import (
     error_distribution,
-    input_prefix_ratio_allclose,
     mapped_pool_ratio_allclose,
     ratio_allclose,
     ratio_reldiff,
-    sampled_ids_golden_compare,
-    stacked_mapped_pool_ratio_allclose,
     topk_pair_compare,
     validate_golden,
 )
@@ -61,13 +58,10 @@ __all__ = [
     "ScalarSpec",
     "validate_golden",
     "ratio_allclose",
-    "input_prefix_ratio_allclose",
     "mapped_pool_ratio_allclose",
-    "stacked_mapped_pool_ratio_allclose",
     "ratio_reldiff",
     "error_distribution",
     "topk_pair_compare",
-    "sampled_ids_golden_compare",
     "RunResult",
     "run",
     "run_jit",
