+++ golden/validation.py
@@ -69,60 +69,6 @@ def _nonfinite_error(actual: torch.Tensor, expected: torch.Tensor) -> str:
     )
 
 
-def _stacked_mapped_nonfinite_coordinates(
-    actual: torch.Tensor,
-    expected: torch.Tensor,
-    *,
-    valid_mapping: torch.Tensor,
-    valid_mapping_items: torch.Tensor,
-    feature_shape: tuple[int, ...],
-    rank: int,
-    layer_index: int,
-    blocks_per_layer: int,
-    block_size: int,
-    max_show: int,
-) -> str:
-    """Format logical and physical coordinates for mapped non-finite values."""
-    nonfinite = ~torch.isfinite(actual) | ~torch.isfinite(expected)
-    coordinates = nonfinite.nonzero(as_tuple=False)
-    total = int(coordinates.shape[0])
-    if total == 0:
-        return ""
-
-    show_count = min(max(max_show, 0), total)
-    lines = [f"    non-finite mapped coordinate(s): showing {show_count}/{total}"]
-    for coordinate in coordinates[:show_count]:
-        mapped_offset = int(coordinate[0].item())
-        flat_feature = int(coordinate[1].item())
-        remaining = flat_feature
-        feature_coordinates = []
-        for size in reversed(feature_shape):
-            feature_coordinates.append(remaining % size)
-            remaining //= size
-        feature_coordinates.reverse()
-
-        mapped_row = int(valid_mapping_items[mapped_offset].item())
-        physical_row = int(valid_mapping[mapped_offset].item())
-        pool_block = layer_index * blocks_per_layer + physical_row // block_size
-        block_row = physical_row % block_size
-        logical_coordinate = (mapped_row, *feature_coordinates)
-        physical_coordinate = (rank, pool_block, block_row, *feature_coordinates)
-        actual_value = actual[mapped_offset, flat_feature].item()
-        expected_value = expected[mapped_offset, flat_feature].item()
-        lines.append(
-            "      "
-            f"logical(mapped_row, feature...)={logical_coordinate} "
-            "physical_pool(rank, block, block_row, feature...)="
-            f"{physical_coordinate} layer_physical_row={physical_row} "
-            f"actual={actual_value!r} expected={expected_value!r}"
-        )
-    if total > show_count:
-        lines.append(
-            f"      ... and {total - show_count} more non-finite coordinate(s)"
-        )
-    return "\n".join(lines)
-
-
 def validate_golden(
     outputs: dict[str, torch.Tensor],
     golden: dict[str, torch.Tensor],
@@ -360,109 +306,6 @@ def topk_pair_compare(
     return cmp
 
 
-def sampled_ids_golden_compare(
-    *,
-    logits_name: str = "logits",
-    row_indices_name: str = "logit_row_indices",
-    sampled_id_column: int = 0,
-    max_show: int = 20,
-) -> Callable:
-    """Validate sampled IDs against golden and the device's own logits.
-
-    This combines two independent contracts on every selected logit row:
-
-    * semantic correctness — the complete sampled-ID tensor equals golden;
-    * sampler self-consistency — the sampled token equals ``argmax`` of the
-      actual device logits.
-
-    Checking only the second condition can report PASS for a confidently
-    wrong model token, so full-model validation should use both.
-    """
-
-    def compare(
-        actual: torch.Tensor,
-        expected: torch.Tensor,
-        *,
-        actual_outputs: dict[str, torch.Tensor],
-        expected_outputs: dict[str, torch.Tensor],
-        inputs: dict[str, torch.Tensor],
-        rtol: float,
-        atol: float,
-    ) -> tuple[bool, str]:
-        del expected_outputs, rtol, atol
-        if actual.shape != expected.shape:
-            return False, (
-                f"    sampled_ids shape mismatch: actual={tuple(actual.shape)} "
-                f"expected={tuple(expected.shape)}"
-            )
-        if actual.ndim < 1 or not 0 <= sampled_id_column < actual.shape[-1]:
-            return False, (
-                f"    sampled_id_column={sampled_id_column} out of range for "
-                f"shape {tuple(actual.shape)}"
-            )
-        if logits_name not in actual_outputs:
-            return False, (
-                f"    compare_fn misconfigured: missing actual output '{logits_name}'"
-            )
-        if row_indices_name not in inputs:
-            return False, (
-                f"    compare_fn misconfigured: missing input '{row_indices_name}'"
-            )
-
-        logits = actual_outputs[logits_name].cpu()
-        row_indices = inputs[row_indices_name].cpu()
-        leading_shape = actual.shape[:-1]
-        if tuple(logits.shape[:-1]) != tuple(leading_shape):
-            return False, (
-                f"    '{logits_name}' leading shape must be {tuple(leading_shape)}, "
-                f"got {tuple(logits.shape)}"
-            )
-        if tuple(row_indices.shape) != tuple(leading_shape):
-            return False, (
-                f"    '{row_indices_name}' shape must be {tuple(leading_shape)}, "
-                f"got {tuple(row_indices.shape)}"
-            )
-
-        actual_cpu = actual.cpu()
-        expected_cpu = expected.cpu()
-        valid = row_indices >= 0
-        device_argmax = torch.argmax(logits, dim=-1).to(actual_cpu.dtype)
-        actual_ids = actual_cpu[..., sampled_id_column]
-        expected_ids = expected_cpu[..., sampled_id_column]
-        semantic_bad = actual_cpu != expected_cpu
-        self_bad = valid & (actual_ids != device_argmax)
-        failures: list[str] = []
-
-        if semantic_bad.any().item():
-            failures.append(
-                f"    sampled_ids differ from golden: "
-                f"bad={int(semantic_bad.count_nonzero().item())}/{actual_cpu.numel()}"
-            )
-        if self_bad.any().item():
-            failures.append(
-                f"    sampled_ids disagree with argmax(actual {logits_name}): "
-                f"bad={int(self_bad.count_nonzero().item())}/{int(valid.count_nonzero().item())}"
-            )
-
-        interesting = valid & (
-            (actual_ids != expected_ids) | (actual_ids != device_argmax)
-        )
-        coords = interesting.nonzero(as_tuple=False)
-        for coord in coords[:max_show]:
-            index = tuple(int(value.item()) for value in coord)
-            failures.append(
-                f"      row={index} actual={int(actual_ids[index].item())} "
-                f"golden={int(expected_ids[index].item())} "
-                f"device_argmax={int(device_argmax[index].item())}"
-            )
-        if coords.shape[0] > max_show:
-            failures.append(f"      ... and {coords.shape[0] - max_show} more")
-        return (not failures), "\n".join(failures)
-
-    compare.__name__ = "sampled_ids_golden_compare"
-    return compare
-
-
 def ratio_allclose(
     atol: float | None = None,
     rtol: float | None = None,
@@ -594,106 +437,6 @@ def ratio_allclose(
         f"valid_rows={valid_rows}, valid_axis={valid_axis}, zero_tail={zero_tail})"
     )
     return cmp
-
-
-def input_prefix_ratio_allclose(
-    valid_rows_name: str,
-    *,
-    valid_axis: int = 0,
-    exact_tail: bool = True,
-    atol: float | None = None,
-    rtol: float | None = None,
-    max_error_ratio: float = 0.005,
-    max_show: int = 10,
-) -> Callable:
-    """Compare an active prefix whose row count comes from a scalar input.
-
-    Unlike :func:`ratio_allclose`'s static ``valid_rows`` option, this helper
-    resolves the prefix length from ``inputs[valid_rows_name]`` for every
-    validation call.  This is important for golden-data replay: cached scalar
-    inputs override the initializer used to rebuild the specs.
-
-    When ``exact_tail`` is true, rows outside the active prefix must remain
-    exactly equal to the golden output.  The inactive rows therefore cannot
-    dilute the active-region error ratio or hide a stray write.
-    """
-    prefix_compare = ratio_allclose(
-        atol=atol,
-        rtol=rtol,
-        max_error_ratio=max_error_ratio,
-        max_show=max_show,
-    )
-
-    def compare(
-        actual: torch.Tensor,
-        expected: torch.Tensor,
-        *,
-        actual_outputs: dict[str, torch.Tensor],
-        expected_outputs: dict[str, torch.Tensor],
-        inputs: dict[str, torch.Tensor],
-        rtol: float,
-        atol: float,
-    ) -> tuple[bool, str]:
-        if actual.shape != expected.shape:
-            return False, (
-                f"    shape mismatch: actual={tuple(actual.shape)} "
-                f"expected={tuple(expected.shape)}"
-            )
-        if valid_rows_name not in inputs:
-            return False, (
-                f"    compare_fn misconfigured: missing input '{valid_rows_name}'"
-            )
-
-        ndim = actual.ndim
-        axis = valid_axis if valid_axis >= 0 else valid_axis + ndim
-        if not 0 <= axis < ndim:
-            return False, (
-                f"    valid_axis={valid_axis} out of range for shape {tuple(actual.shape)}"
-            )
-
-        valid_rows_value = inputs[valid_rows_name].cpu()
-        if valid_rows_value.numel() != 1:
-            return False, (
-                f"    '{valid_rows_name}' must be a scalar, "
-                f"got shape {tuple(valid_rows_value.shape)}"
-            )
-        valid_rows = int(valid_rows_value.item())
-        total_rows = actual.shape[axis]
-        if not 0 <= valid_rows <= total_rows:
-            return False, (
-                f"    {valid_rows_name}={valid_rows} out of range for axis "
-                f"{axis} of length {total_rows}"
-            )
-
-        if exact_tail and valid_rows < total_rows:
-            actual_tail = actual.narrow(axis, valid_rows, total_rows - valid_rows)
-            expected_tail = expected.narrow(axis, valid_rows, total_rows - valid_rows)
-            unequal_tail = actual_tail != expected_tail
-            if unequal_tail.any().item():
-                return False, (
-                    f"    inactive tail differs from golden: "
-                    f"changed_values={int(unequal_tail.count_nonzero().item())} "
-                    f"axis={axis} active_rows={valid_rows}"
-                )
-
-        actual_prefix = actual.narrow(axis, 0, valid_rows)
-        expected_prefix = expected.narrow(axis, 0, valid_rows)
-        return prefix_compare(
-            actual_prefix,
-            expected_prefix,
-            actual_outputs=actual_outputs,
-            expected_outputs=expected_outputs,
-            inputs=inputs,
-            rtol=rtol,
-            atol=atol,
-        )
-
-    compare.__name__ = (
-        f"input_prefix_ratio_allclose(valid_rows_name={valid_rows_name}, "
-        f"valid_axis={valid_axis}, exact_tail={exact_tail}, atol={atol}, "
-        f"rtol={rtol}, max_error_ratio={max_error_ratio})"
-    )
-    return compare
 
 
 def mapped_pool_ratio_allclose(
@@ -877,246 +620,6 @@ def mapped_pool_ratio_allclose(
         f"mapped_pool_ratio_allclose(mapping={mapping_name}, shape={mapping_shape}, "
         f"block_size={block_size}, leading_rank_axis={leading_rank_axis}, "
         f"atol={atol}, rtol={rtol}, max_error_ratio={max_error_ratio})"
-    )
-    return compare
-
-
-def stacked_mapped_pool_ratio_allclose(
-    layer_mapping_names: tuple[str | None, ...],
-    *,
-    mapping_shape: tuple[int, int],
-    block_size: int,
-    active_rows_name: str,
-    layer_labels: tuple[int, ...] | None = None,
-    pool_name: str = "pool",
-    atol: float | None = None,
-    rtol: float | None = None,
-    max_error_ratio: float = 0.005,
-    max_show: int = 3,
-) -> Callable:
-    """Compare active mapped rows in a rank- and layer-stacked pool.
-
-    The pool layout is ``[ranks, layers * blocks, block_size, ...]``.  Each
-    entry in ``layer_mapping_names`` selects the rank-local mapping input for
-    that layer; ``None`` means the layer must not write this pool.  Only the
-    leading ``inputs[active_rows_name]`` mapping entries participate, which
-    makes replay honor the cached active-token scalar.
-
-    Mapped values use a ratio-based numerical comparison independently for
-    every layer and rank.  All other physical rows must remain exactly equal
-    to golden.  Failure diagnostics therefore identify the first bad logical
-    layer/rank instead of reporting a ratio diluted by the unused pool.
-    """
-    if not layer_mapping_names:
-        raise ValueError("layer_mapping_names must not be empty")
-    if len(mapping_shape) != 2 or any(dim <= 0 for dim in mapping_shape):
-        raise ValueError(
-            f"mapping_shape must be (ranks, mapped_items), got {mapping_shape}"
-        )
-    if block_size <= 0:
-        raise ValueError(f"block_size must be positive, got {block_size}")
-    if layer_labels is None:
-        layer_labels = tuple(range(len(layer_mapping_names)))
-    if len(layer_labels) != len(layer_mapping_names):
-        raise ValueError(
-            "layer_labels and layer_mapping_names must have the same length, "
-            f"got {len(layer_labels)} and {len(layer_mapping_names)}"
-        )
-
-    mapped_compare = ratio_allclose(
-        atol=atol,
-        rtol=rtol,
-        max_error_ratio=max_error_ratio,
-        max_show=max_show,
-    )
-    integer_dtypes = (
-        torch.int8,
-        torch.int16,
-        torch.int32,
-        torch.int64,
-        torch.uint8,
-    )
-
-    def compare(
-        actual: torch.Tensor,
-        expected: torch.Tensor,
-        *,
-        actual_outputs: dict[str, torch.Tensor],
-        expected_outputs: dict[str, torch.Tensor],
-        inputs: dict[str, torch.Tensor],
-        rtol: float,
-        atol: float,
-    ) -> tuple[bool, str]:
-        if actual.shape != expected.shape:
-            return False, (
-                f"    {pool_name} shape mismatch: actual={tuple(actual.shape)} "
-                f"expected={tuple(expected.shape)}"
-            )
-        rank_count, mapped_items = mapping_shape
-        layer_count = len(layer_mapping_names)
-        if actual.ndim < 4 or actual.shape[0] != rank_count:
-            return False, (
-                f"    expected {pool_name} layout "
-                "[ranks, layers * blocks, block_size, ...] with "
-                f"ranks={rank_count}, got {tuple(actual.shape)}"
-            )
-        if actual.shape[1] % layer_count != 0 or actual.shape[2] != block_size:
-            return False, (
-                f"    expected {pool_name} layout "
-                "[ranks, layers * blocks, block_size, ...] with "
-                f"layers={layer_count} block_size={block_size}, "
-                f"got {tuple(actual.shape)}"
-            )
-        if active_rows_name not in inputs:
-            return False, (
-                f"    compare_fn misconfigured: missing input '{active_rows_name}'"
-            )
-        active_value = inputs[active_rows_name].cpu()
-        if active_value.numel() != 1:
-            return False, (
-                f"    '{active_rows_name}' must be a scalar, "
-                f"got shape {tuple(active_value.shape)}"
-            )
-        active_rows = int(active_value.item())
-        if not 0 <= active_rows <= mapped_items:
-            return False, (
-                f"    {active_rows_name}={active_rows} out of range for "
-                f"mapping length {mapped_items}"
-            )
-
-        mapping_cache: dict[str, torch.Tensor] = {}
-        for mapping_name in set(layer_mapping_names) - {None}:
-            mapping = inputs.get(mapping_name)
-            if mapping is None:
-                return False, (
-                    f"    compare_fn misconfigured: missing input '{mapping_name}'"
-                )
-            if mapping.dtype not in integer_dtypes:
-                return False, (
-                    f"    '{mapping_name}' must have an integer dtype, "
-                    f"got {mapping.dtype}"
-                )
-            if tuple(mapping.shape) != mapping_shape:
-                return False, (
-                    f"    '{mapping_name}' must have shape {mapping_shape}, "
-                    f"got {tuple(mapping.shape)}"
-                )
-            mapping_cache[mapping_name] = mapping.cpu().to(torch.int64)
-
-        blocks_per_layer = actual.shape[1] // layer_count
-        rows_per_layer = blocks_per_layer * block_size
-        actual_rows = actual.cpu().reshape(
-            rank_count, layer_count, rows_per_layer, -1
-        )
-        expected_rows = expected.cpu().reshape(
-            rank_count, layer_count, rows_per_layer, -1
-        )
-        failures: list[str] = []
-
-        for layer_index, (layer_label, mapping_name) in enumerate(
-            zip(layer_labels, layer_mapping_names)
-        ):
-            for rank in range(rank_count):
-                if mapping_name is None:
-                    mapping = torch.empty(0, dtype=torch.int64)
-                else:
-                    mapping = mapping_cache[mapping_name][rank, :active_rows]
-
-                invalid_negative = mapping < -1
-                if invalid_negative.any().item():
-                    item = int(invalid_negative.nonzero(as_tuple=False)[0, 0].item())
-                    failures.append(
-                        f"    layer={layer_label} rank={rank} mapping='{mapping_name}' "
-                        f"item={item} value={int(mapping[item].item())}: "
-                        "only -1 is a negative sentinel"
-                    )
-                    continue
-                valid_mapping_items = (mapping >= 0).nonzero(as_tuple=False).flatten()
-                valid_mapping = mapping[valid_mapping_items]
-                if (valid_mapping >= rows_per_layer).any().item():
-                    value = int(valid_mapping[valid_mapping >= rows_per_layer][0].item())
-                    failures.append(
-                        f"    layer={layer_label} rank={rank} mapping='{mapping_name}' "
-                        f"row={value} outside physical row range "
-                        f"[0, {rows_per_layer})"
-                    )
-                    continue
-                if valid_mapping.numel() > 1:
-                    unique_rows, counts = torch.unique(
-                        valid_mapping, return_counts=True
-                    )
-                    duplicates = counts > 1
-                    if duplicates.any().item():
-                        duplicate_row = int(unique_rows[duplicates][0].item())
-                        failures.append(
-                            f"    layer={layer_label} rank={rank} "
-                            f"mapping='{mapping_name}' contains duplicate "
-                            f"physical row {duplicate_row}"
-                        )
-                        continue
-
-                written_rows = torch.zeros(rows_per_layer, dtype=torch.bool)
-                written_rows[valid_mapping] = True
-                actual_layer = actual_rows[rank, layer_index]
-                expected_layer = expected_rows[rank, layer_index]
-                stray_values = actual_layer[~written_rows] != expected_layer[~written_rows]
-                if stray_values.any().item():
-                    changed_rows = (
-                        (actual_layer[~written_rows] != expected_layer[~written_rows])
-                        .any(dim=-1)
-                        .count_nonzero()
-                        .item()
-                    )
-                    failures.append(
-                        f"    layer={layer_label} rank={rank} mapping='{mapping_name}' "
-                        f"has writes outside active mapped rows: "
-                        f"changed_rows={int(changed_rows)} "
-                        f"changed_values={int(stray_values.count_nonzero().item())}"
-                    )
-                    continue
-                if valid_mapping.numel() == 0:
-                    continue
-
-                mapped_actual = actual_layer[valid_mapping]
-                mapped_expected = expected_layer[valid_mapping]
-                ok, detail = mapped_compare(
-                    mapped_actual,
-                    mapped_expected,
-                    actual_outputs=actual_outputs,
-                    expected_outputs=expected_outputs,
-                    inputs=inputs,
-                    rtol=rtol,
-                    atol=atol,
-                )
-                if not ok:
-                    nonfinite_coordinates = _stacked_mapped_nonfinite_coordinates(
-                        mapped_actual,
-                        mapped_expected,
-                        valid_mapping=valid_mapping,
-                        valid_mapping_items=valid_mapping_items,
-                        feature_shape=tuple(actual.shape[3:]),
-                        rank=rank,
-                        layer_index=layer_index,
-                        blocks_per_layer=blocks_per_layer,
-                        block_size=block_size,
-                        max_show=max_show,
-                    )
-                    coordinate_detail = (
-                        f"\n{nonfinite_coordinates}" if nonfinite_coordinates else ""
-                    )
-                    failures.append(
-                        f"    layer={layer_label} rank={rank} "
-                        f"mapping='{mapping_name}' mapped_rows={valid_mapping.numel()}\n"
-                        f"{detail}{coordinate_detail}"
-                    )
-
-        return (not failures), "\n".join(failures)
-
-    compare.__name__ = (
-        f"stacked_mapped_pool_ratio_allclose(pool={pool_name}, "
-        f"layers={len(layer_mapping_names)}, block_size={block_size}, "
-        f"active_rows_name={active_rows_name}, atol={atol}, rtol={rtol}, "
-        f"max_error_ratio={max_error_ratio})"
     )
     return compare
