+++ models/deepseek_v4_pro/prefill_layer.py
@@ -1729,7 +1729,7 @@ if __name__ == "__main__":
         start_positions=start_positions,
     )
     if args.weights is not None:
-        from weights_flash import apply_real_layer_weights
+        from utils.weights_flash import apply_real_layer_weights
 
         count = apply_real_layer_weights(specs, args.weights, layer_id=args.layer_id, ep=N_RANKS)
         print(f"[RUN] real weights: layer {args.layer_id}, {count} tensors from {args.weights}", flush=True)
